requirements.txt
flask==3.1.0
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
werkzeug==3.1.3
gunicorn==23.0.0
